Add files required for IG publishing
This is just a snapshot of the files. The IG Publisher and the publishing process will keep updating these files with each published version of any of the implementation guides.

These files are stored here for reference, so that it's easier to get things back on track if something goes wrong in the production environment.