Add a simple directory listing page
Also add links to HL7 Finland's implementation guides published in Simplifier.

And add plenty of resources too. Mainly so that they are available if we want to use them in the future. Note that not all resources used by bootstrap are included. And some resources may be in wrong places, the references from css files and scripts have not been extensively checked.